Look in override file to check for referenced apps (#2184)
Search the override for for `.env.app.appname` strings, and if found, consider `appname` to be referenced.

Delete any `.env.app.appname` files for apps that are no longer referenced, i.e., do not have any `APPNAME__*` variables in the global `.env` file, do not have any variables in the `.env.app.appname` file, and do not have `.env.app.appname` in the override file.

+++ .scripts/env_update.sh
@@ -10,9 +10,24 @@ env_update() {
     #fi
     notice "Updating environment variable files."
 
-    local -l applist
-    applist="$(run_script 'app_list_referenced')"
+    local -al applist
+    readarray -t applist < <(
+        run_script 'app_list_hasvarfile'
+    )
+    for appname in "${applist[@]}"; do
+        if ! run_script 'app_is_referenced' "${appname}"; then
+            local AppEnvFile
+            AppEnvFile="$(run_script 'app_env_file' "${appname}")"
+            run_script 'set_permissions' "${AppEnvFile}"
+            notice "Deleting '${C["File"]}${AppEnvFile}${NC}'."
+            rm -f "${AppEnvFile}" ||
+                warn "Failed to remove '${C["File"]}${AppEnvFile}${NC}'.\nFailing command: ${C["FailingCommand"]}rm -f \"${AppEnvFile}\""
+        fi
+    done
 
+    readarray -t applist < <(
+        run_script 'app_list_referenced'
+    )
     # Format the global .env file
     if ! run_script 'needs_env_update' "${COMPOSE_ENV}"; then
         info "'${C["File"]}${COMPOSE_ENV}'${NC} already updated."
@@ -27,17 +42,19 @@ env_update() {
             run_script 'env_format_lines' "${ENV_LINES_FILE}" "${COMPOSE_ENV_DEFAULT_FILE}" ""
         )
 
-        for appname in ${applist}; do
-            local APP_DEFAULT_GLOBAL_ENV_FILE=""
-            local -a UPDATED_APP_ENV_LINES=()
-            if ! run_script 'app_is_user_defined' "${appname}"; then
-                APP_DEFAULT_GLOBAL_ENV_FILE="$(run_script 'app_instance_file' "${appname}" ".env")"
-            fi
-            run_script 'appvars_lines' "${appname}" > "${ENV_LINES_FILE}"
-            readarray -t -O ${#UPDATED_ENV_LINES[@]} UPDATED_ENV_LINES < <(
-                run_script 'env_format_lines' "${ENV_LINES_FILE}" "${APP_DEFAULT_GLOBAL_ENV_FILE}" "${appname}"
-            )
-        done
+        if [[ -n ${applist[*]-} ]]; then
+            for appname in "${applist[@]}"; do
+                local APP_DEFAULT_GLOBAL_ENV_FILE=""
+                local -a UPDATED_APP_ENV_LINES=()
+                if ! run_script 'app_is_user_defined' "${appname}"; then
+                    APP_DEFAULT_GLOBAL_ENV_FILE="$(run_script 'app_instance_file' "${appname}" ".env")"
+                fi
+                run_script 'appvars_lines' "${appname}" > "${ENV_LINES_FILE}"
+                readarray -t -O ${#UPDATED_ENV_LINES[@]} UPDATED_ENV_LINES < <(
+                    run_script 'env_format_lines' "${ENV_LINES_FILE}" "${APP_DEFAULT_GLOBAL_ENV_FILE}" "${appname}"
+                )
+            done
+        fi
         rm -f "${ENV_LINES_FILE}" ||
             warn "Failed to remove temporary '${C["File"]}.env${NC}' update file.\nFailing command: ${C["FailingCommand"]}rm -f \"${ENV_LINES_FILE}\""
 
@@ -54,34 +71,40 @@ env_update() {
     fi
 
     # Process all referenced .env.app.appname files
-    for appname in ${applist}; do
-        local APP_ENV_FILE
-        APP_ENV_FILE="$(run_script 'app_env_file' "${appname}")"
-        if ! run_script 'needs_env_update' "${APP_ENV_FILE}"; then
-            info "'${C["File"]}${APP_ENV_FILE}'${NC} already updated."
-        else
-            notice "Updating '${C["File"]}${APP_ENV_FILE}${NC}'."
-            local APP_DEFAULT_ENV_FILE=""
-            if ! run_script 'app_is_user_defined' "${appname}"; then
-                APP_DEFAULT_ENV_FILE="$(run_script 'app_instance_file' "${appname}" ".env.app.*")"
+    if [[ -n ${applist[*]-} ]]; then
+        for appname in "${applist[@]-}"; do
+            local APP_ENV_FILE
+            APP_ENV_FILE="$(run_script 'app_env_file' "${appname}")"
+            if ! run_script 'needs_env_update' "${APP_ENV_FILE}"; then
+                info "'${C["File"]}${APP_ENV_FILE}'${NC} already updated."
+            else
+                if [[ ! -f ${APP_ENV_FILE} ]]; then
+                    notice "Creating '${C["File"]}${APP_ENV_FILE}${NC}'."
+                else
+                    notice "Updating '${C["File"]}${APP_ENV_FILE}${NC}'."
+                fi
+                local APP_DEFAULT_ENV_FILE=""
+                if ! run_script 'app_is_user_defined' "${appname}"; then
+                    APP_DEFAULT_ENV_FILE="$(run_script 'app_instance_file' "${appname}" ".env.app.*")"
+                fi
+                local -a UPDATED_APP_ENV_LINES=()
+                readarray -t UPDATED_APP_ENV_LINES < <(
+                    run_script 'env_format_lines' "${APP_ENV_FILE}" "${APP_DEFAULT_ENV_FILE}" "${appname}"
+                )
+                local MKTEMP_APP_ENV_UPDATED
+                MKTEMP_APP_ENV_UPDATED=$(mktemp -t "${APPLICATION_NAME}.${FUNCNAME[0]}.MKTEMP_APP_ENV_UPDATED.XXXXXXXXXX") ||
+                    fatal "Failed to create temporary update '${C["File"]}.env.app.${appname}${NC}' file.\nFailing command: ${C["FailingCommand"]}mktemp -t \"${APPLICATION_NAME}.${FUNCNAME[0]}.MKTEMP_APP_ENV_UPDATED.XXXXXXXXXX\"${NC}"
+                printf '%s\n' "${UPDATED_APP_ENV_LINES[@]}" > "${MKTEMP_APP_ENV_UPDATED}" ||
+                    fatal "Failed to write temporary '${C["File"]}.env.app.${appname}${NC}' update file."
+                cp -f "${MKTEMP_APP_ENV_UPDATED}" "${APP_ENV_FILE}" ||
+                    fatal "Failed to copy file.\nFailing command: ${C["FailingCommand"]}cp -f \"${MKTEMP_APP_ENV_UPDATED}\" \"${APP_ENV_FILE}\""
+                rm -f "${MKTEMP_APP_ENV_UPDATED}" ||
+                    warn "Failed to remove temporary '${C["File"]}.env.app.${appname}${NC}' update file.\nFailing command: ${C["FailingCommand"]}rm -f \"${MKTEMP_APP_ENV_UPDATED}\""
+                run_script 'set_permissions' "${APP_ENV_FILE}"
+                #run_script 'unset_needs_env_update' "${APP_ENV_FILE}"
             fi
-            local -a UPDATED_APP_ENV_LINES=()
-            readarray -t UPDATED_APP_ENV_LINES < <(
-                run_script 'env_format_lines' "${APP_ENV_FILE}" "${APP_DEFAULT_ENV_FILE}" "${appname}"
-            )
-            local MKTEMP_APP_ENV_UPDATED
-            MKTEMP_APP_ENV_UPDATED=$(mktemp -t "${APPLICATION_NAME}.${FUNCNAME[0]}.MKTEMP_APP_ENV_UPDATED.XXXXXXXXXX") ||
-                fatal "Failed to create temporary update '${C["File"]}.env.app.${appname}${NC}' file.\nFailing command: ${C["FailingCommand"]}mktemp -t \"${APPLICATION_NAME}.${FUNCNAME[0]}.MKTEMP_APP_ENV_UPDATED.XXXXXXXXXX\"${NC}"
-            printf '%s\n' "${UPDATED_APP_ENV_LINES[@]}" > "${MKTEMP_APP_ENV_UPDATED}" ||
-                fatal "Failed to write temporary '${C["File"]}.env.app.${appname}${NC}' update file."
-            cp -f "${MKTEMP_APP_ENV_UPDATED}" "${APP_ENV_FILE}" ||
-                fatal "Failed to copy file.\nFailing command: ${C["FailingCommand"]}cp -f \"${MKTEMP_APP_ENV_UPDATED}\" \"${APP_ENV_FILE}\""
-            rm -f "${MKTEMP_APP_ENV_UPDATED}" ||
-                warn "Failed to remove temporary '${C["File"]}.env.app.${appname}${NC}' update file.\nFailing command: ${C["FailingCommand"]}rm -f \"${MKTEMP_APP_ENV_UPDATED}\""
-            run_script 'set_permissions' "${APP_ENV_FILE}"
-            #run_script 'unset_needs_env_update' "${APP_ENV_FILE}"
-        fi
-    done
+        done
+    fi
 
     #run_script 'env_sanitize'
     run_script 'unset_needs_env_update'
